feat: implement document upload and metadata management in onboarding process
- Added a new endpoint for uploading documents with a maximum size limit of 10MB.
- Introduced a new form for users to upload multiple documents and display selected files.
- Implemented functionality to handle file uploads, including validation for file size and type.
- Created a new method to update document metadata in the database.
- Enhanced the onboarding documents page to show uploaded documents and allow users to update their metadata.
- Updated the database schema to include the correct schema for the documents table.
- Refactored existing code to improve readability and maintainability.

+++ pkg/types/document.go
@@ -9,7 +9,7 @@ type NeedDocument struct {
 	UserID        string    `db:"user_id" json:"userId"`
 	DocumentType  string    `db:"document_type" json:"documentType"`
 	FileName      string    `db:"file_name" json:"fileName"`
-	FileSizeBytes int       `db:"file_size_bytes" json:"fileSizeBytes"`
+	FileSizeBytes int64     `db:"file_size_bytes" json:"fileSizeBytes"`
 	MimeType      string    `db:"mime_type" json:"mimeType"`
 	StorageKey    string    `db:"storage_key" json:"storageKey"`
 	UploadedAt    time.Time `db:"uploaded_at" json:"uploadedAt"`
@@ -17,10 +17,10 @@ type NeedDocument struct {
 
 // Document type constants
 const (
-	DocTypeID                  = "id"
-	DocTypeUtilityBill         = "utility_bill"
-	DocTypeMedicalRecord       = "medical_record"
-	DocTypeIncomeVerification  = "income_verification"
-	DocTypeEvictionNotice      = "eviction_notice"
-	DocTypeOther               = "other"
+	DocTypeID                 = "id"
+	DocTypeUtilityBill        = "utility_bill"
+	DocTypeMedicalRecord      = "medical_record"
+	DocTypeIncomeVerification = "income_verification"
+	DocTypeEvictionNotice     = "eviction_notice"
+	DocTypeOther              = "other"
 )
